test(content): split PageImageContentAnalyzerTest by content type
Split the oversized analyzer suite into image-block and HTML-snippet
classes with a shared test case, matching the B6–B8 hotspot pattern.
