+++ tests/Content/Integration/Page/PageImageContentAnalyzerImageBlockTest.php
@@ -4,24 +4,12 @@ declare(strict_types=1);
 
 namespace App\Tests\Content\Integration\Page;
 
-use App\Content\Application\Page\PageImageContentAnalyzer;
 use App\Content\Infrastructure\Factory\MediaFactory;
 use App\Content\Infrastructure\Factory\PageFactory;
-use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use Symfony\Component\Filesystem\Path;
-use Zenstruck\Foundry\Attribute\ResetDatabase;
-use Zenstruck\Foundry\Test\Factories;
 
-#[ResetDatabase]
-final class PageImageContentAnalyzerTest extends KernelTestCase
+final class PageImageContentAnalyzerImageBlockTest extends PageImageContentAnalyzerTestCase
 {
-    use Factories;
-
-    protected function setUp(): void
-    {
-        self::bootKernel();
-    }
-
     public function testRecommendsMigratingSmallFullWidthImageToAuto(): void
     {
         $media = MediaFactory::createOne([
@@ -124,28 +112,6 @@ final class PageImageContentAnalyzerTest extends KernelTestCase
         self::assertSame('Sync media files locally', $findings[0]->recommendation);
     }
 
-    public function testDetectsRichtextSnippetWithLegacyLargeSizeClass(): void
-    {
-        $page = PageFactory::createOne([
-            'slug' => 'analyzer-richtext-snippet',
-            'content' => [
-                [
-                    'type' => 'richtext',
-                    'enabled' => true,
-                    'data' => [
-                        'html' => '<figure class="page-content-image page-content-image--size-lg"><img src="/uploads/media/snippet.png" alt="Snippet"></figure>',
-                    ],
-                ],
-            ],
-        ]);
-
-        $findings = $this->analyzer()->analyze($page->getId());
-
-        self::assertCount(1, $findings);
-        self::assertStringContainsString('richtext', $findings[0]->blockType);
-        self::assertSame('lg', $findings[0]->currentSize);
-    }
-
     public function testAnalyzeAllPagesWhenNoPageIdFilter(): void
     {
         $this->createImagePage('analyzer-all-1');
@@ -159,88 +125,6 @@ final class PageImageContentAnalyzerTest extends KernelTestCase
     public function testReturnsEmptyFindingsForUnknownPageId(): void
     {
         self::assertSame([], $this->analyzer()->analyze(999_999));
-    }
-
-    public function testDetectsHighlightBlockWithFloatedSnippet(): void
-    {
-        $media = MediaFactory::createOne(['filename' => 'highlight-float.png']);
-
-        $page = PageFactory::createOne([
-            'slug' => 'analyzer-highlight',
-            'content' => [
-                [
-                    'type' => 'highlight',
-                    'enabled' => true,
-                    'data' => [
-                        'html' => sprintf(
-                            '<figure class="page-content-image page-content-image--size-md page-content-image--float-right"><img src="/uploads/media/%s" alt="Highlight"></figure>',
-                            $media->getFilename(),
-                        ),
-                    ],
-                ],
-            ],
-        ]);
-
-        $findings = $this->analyzer()->analyze($page->getId());
-
-        self::assertCount(1, $findings);
-        self::assertStringContainsString('highlight', $findings[0]->blockType);
-        self::assertSame('right', $findings[0]->float);
-        self::assertSame('md', $findings[0]->currentSize);
-        self::assertSame('No change', $findings[0]->recommendation);
-    }
-
-    public function testDetectsAccordionItemWithImageSnippet(): void
-    {
-        $page = PageFactory::createOne([
-            'slug' => 'analyzer-accordion',
-            'content' => [
-                [
-                    'type' => 'accordion',
-                    'enabled' => true,
-                    'data' => [
-                        'items' => [
-                            [
-                                'title' => 'FAQ',
-                                'html' => '<figure class="page-content-image page-content-image--size-lg"><img src="/uploads/media/accordion.png" alt="Accordion"></figure>',
-                            ],
-                        ],
-                    ],
-                ],
-            ],
-        ]);
-
-        $findings = $this->analyzer()->analyze($page->getId());
-
-        self::assertCount(1, $findings);
-        self::assertStringContainsString('accordion item 1', $findings[0]->blockType);
-    }
-
-    public function testDetectsInlineImageWithoutFigureClasses(): void
-    {
-        $media = MediaFactory::createOne(['filename' => 'inline-only.png']);
-
-        $page = PageFactory::createOne([
-            'slug' => 'analyzer-inline-img',
-            'content' => [
-                [
-                    'type' => 'richtext',
-                    'enabled' => true,
-                    'data' => [
-                        'html' => sprintf(
-                            '<p>Text <img src="/uploads/media/%s" alt="Inline"></p>',
-                            $media->getFilename(),
-                        ),
-                    ],
-                ],
-            ],
-        ]);
-
-        $findings = $this->analyzer()->analyze($page->getId());
-
-        self::assertCount(1, $findings);
-        self::assertStringContainsString('inline img', $findings[0]->blockType);
-        self::assertSame('auto', $findings[0]->currentSize);
     }
 
     public function testReportsDimensionsUnknownWhenDatabaseMetadataMissing(): void
@@ -467,75 +351,6 @@ final class PageImageContentAnalyzerTest extends KernelTestCase
         self::assertSame('Review layout', $findings[0]->recommendation);
     }
 
-    public function testUnknownSnippetSizeRecommendsReviewLayout(): void
-    {
-        $media = MediaFactory::createOne(['filename' => 'xl-snippet.png']);
-
-        $page = PageFactory::createOne([
-            'slug' => 'analyzer-unknown-snippet-size',
-            'content' => [
-                [
-                    'type' => 'richtext',
-                    'enabled' => true,
-                    'data' => [
-                        'html' => sprintf(
-                            '<figure class="page-content-image page-content-image--size-xl"><img src="/uploads/media/%s" alt="XL"></figure>',
-                            $media->getFilename(),
-                        ),
-                    ],
-                ],
-            ],
-        ]);
-
-        $findings = $this->analyzer()->analyze($page->getId());
-
-        self::assertCount(1, $findings);
-        self::assertSame('xl', $findings[0]->currentSize);
-        self::assertSame('Review layout', $findings[0]->recommendation);
-    }
-
-    public function testSnippetWithoutSrcReportsMissingSrc(): void
-    {
-        $page = PageFactory::createOne([
-            'slug' => 'analyzer-snippet-no-src',
-            'content' => [
-                [
-                    'type' => 'richtext',
-                    'enabled' => true,
-                    'data' => [
-                        'html' => '<figure class="page-content-image page-content-image--size-lg"></figure>',
-                    ],
-                ],
-            ],
-        ]);
-
-        $findings = $this->analyzer()->analyze($page->getId());
-
-        self::assertCount(1, $findings);
-        self::assertSame('missing_src', $findings[0]->status);
-    }
-
-    public function testDetectsFloatedLeftSnippetInHtml(): void
-    {
-        $page = PageFactory::createOne([
-            'slug' => 'analyzer-float-left-snippet',
-            'content' => [
-                [
-                    'type' => 'richtext',
-                    'enabled' => true,
-                    'data' => [
-                        'html' => '<figure class="page-content-image page-content-image--size-sm page-content-image--float-left"><img src="/uploads/media/float-left.png" alt="Left"></figure>',
-                    ],
-                ],
-            ],
-        ]);
-
-        $findings = $this->analyzer()->analyze($page->getId());
-
-        self::assertCount(1, $findings);
-        self::assertSame('left', $findings[0]->float);
-    }
-
     public function testResolvesMediaBySrcFilenameWhenMediaIdMissing(): void
     {
         $media = MediaFactory::createOne([
@@ -600,54 +415,6 @@ final class PageImageContentAnalyzerTest extends KernelTestCase
         self::assertSame(1, $findings[0]->imageWidth);
     }
 
-    public function testSkipsAccordionItemsWithoutHtml(): void
-    {
-        $page = PageFactory::createOne([
-            'slug' => 'analyzer-accordion-invalid-item',
-            'content' => [
-                [
-                    'type' => 'accordion',
-                    'enabled' => true,
-                    'data' => [
-                        'items' => [
-                            ['title' => 'No html'],
-                            [
-                                'title' => 'With snippet',
-                                'html' => '<figure class="page-content-image page-content-image--size-md"><img src="/uploads/media/acc-valid.png" alt="Valid"></figure>',
-                            ],
-                        ],
-                    ],
-                ],
-            ],
-        ]);
-
-        $findings = $this->analyzer()->analyze($page->getId());
-
-        self::assertCount(1, $findings);
-        self::assertStringContainsString('accordion item 2', $findings[0]->blockType);
-    }
-
-    public function testSkipsHtmlBlocksWithEmptyContent(): void
-    {
-        $page = PageFactory::createOne([
-            'slug' => 'analyzer-empty-html',
-            'content' => [
-                [
-                    'type' => 'richtext',
-                    'enabled' => true,
-                    'data' => ['html' => ''],
-                ],
-                [
-                    'type' => 'headline',
-                    'enabled' => true,
-                    'data' => ['text' => 'Title only'],
-                ],
-            ],
-        ]);
-
-        self::assertSame([], $this->analyzer()->analyze($page->getId()));
-    }
-
     public function testUnreadableLocalFileReportsDimensionsUnknown(): void
     {
         $media = MediaFactory::createOne([
@@ -708,10 +475,5 @@ final class PageImageContentAnalyzerTest extends KernelTestCase
                 ],
             ],
         ]);
-    }
-
-    private function analyzer(): PageImageContentAnalyzer
-    {
-        return self::getContainer()->get(PageImageContentAnalyzer::class);
     }
 }
